Se mueve lógica de etapa3 a diccionario
